requirements.txt
aiofiles==24.1.0
asgiref==3.8.1
bcrypt==4.2.1
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
dnspython==2.6.1
ecdsa==0.19.0
email-validator==2.2.0
fastapi==0.68.0
h11==0.14.0
idna==3.10
motor==2.5.1
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==1.8.2
pymongo==3.13.0
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
starlette==0.14.2
typing-extensions==4.12.2
uvicorn==0.15.0
